feat: add beaconRequired button to test app
