+++ example/lib/home_page.dart
@@ -1,6 +1,7 @@
 import 'package:drivekit_example/widgets/initialize_listeners.dart';
 import 'package:drivekit_example/widgets/sliver_add_vehicle.dart';
 import 'package:drivekit_example/widgets/sliver_auto_start.dart';
+import 'package:drivekit_example/widgets/sliver_beacons.dart';
 import 'package:drivekit_example/widgets/sliver_crash_detection.dart';
 import 'package:drivekit_example/widgets/sliver_delete_account.dart';
 import 'package:drivekit_example/widgets/sliver_get_api_key.dart';
@@ -14,7 +15,6 @@ import 'package:drivekit_example/widgets/sliver_remove_all_drivekit_listeners.da
 import 'package:drivekit_example/widgets/sliver_remove_all_trip_listeners.dart';
 import 'package:drivekit_example/widgets/sliver_reset.dart';
 import 'package:drivekit_example/widgets/sliver_set_api_key.dart';
-import 'package:drivekit_example/widgets/sliver_set_beacons.dart';
 import 'package:drivekit_example/widgets/sliver_stop_timeout.dart';
 import 'package:drivekit_example/widgets/sliver_trip_lifecycle.dart';
 import 'package:drivekit_example/widgets/sliver_trip_metadata.dart';
@@ -111,7 +111,7 @@ class HomePage extends StatelessWidget {
             SliverGap(32),
             SliverPadding(
               padding: EdgeInsets.symmetric(horizontal: 16),
-              sliver: SliverSetBeacons(),
+              sliver: SliverBeacons(),
             ),
             SliverGap(32),
             SliverPadding(
